feat: add error form_type
